+++ tenmo-server/src/main/java/com/techelevator/tenmo/dao/JdbcTransfer.java
@@ -25,7 +25,7 @@ public class JdbcTransfer implements TransferDao {
     }
 
     @Override
-    public Transfer createTransfer(double amount, int accountFrom, int accountTo, int transferType, int transferStatus) {
+    public Transfer createTransfer(BigDecimal amount, int accountFrom, int accountTo, int transferType, int transferStatus) {
         Transfer transfer = new Transfer(amount, accountFrom, accountTo);
         String sql = "INSERT INTO transfer (transfer_type_id, transfer_status_id, account_from, account_to, amount)" +
                 "VALUES (?,?,?,?,?)";
